\name{BSE-internal}

%BSE.example.R
\alias{BSE.example}

\title{Internal BSE functions}

\description{
  Internal BSE functions. These are generally not to be called by the user.
}

\author{Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
