% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.fishery.estimates.BET.old.f.R
\name{call.fishery.estimates.BET.old.f}
\alias{call.fishery.estimates.BET.old.f}
\title{xxx}
\usage{
call.fishery.estimates.BET.old.f(unlds.bystrat, iyear, PS)
}
\description{
\code{call.fishery.estimates.BET.old.f} yyy
}
