% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.fishery.estimates.f.R
\name{call.fishery.estimates.f}
\alias{call.fishery.estimates.f}
\title{xxx}
\usage{
call.fishery.estimates.f(
  strat.ests.withsamps,
  unlds.bystrat,
  iyear,
  PS,
  Species
)
}
\description{
\code{call.fishery.estimates.f} yyy
}
