% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.stratum.estimates.f.R
\name{call.stratum.estimates.f}
\alias{call.stratum.estimates.f}
\title{xxx}
\usage{
call.stratum.estimates.f(
  unloads.bystrat,
  well.estimates,
  well.strats,
  min.sampsize
)
}
\description{
\code{call.stratum.estimates.f} yyy
}
