% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.strat.flg.f.R
\name{create.strat.flg.f}
\alias{create.strat.flg.f}
\title{xxx}
\usage{
create.strat.flg.f(
  lat.5deg,
  lon.5deg,
  is.lwrght,
  month,
  setype,
  vessel.class,
  PS,
  Species
)
}
\description{
\code{create.strat.flg.f} yyy
}
