% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.strat.flg.f.SKJ.R
\name{create.strat.flg.f.SKJ}
\alias{create.strat.flg.f.SKJ}
\title{xxx}
\usage{
create.strat.flg.f.SKJ(
  lat.5deg,
  lon.5deg,
  is.lwrght,
  month,
  setype,
  vessel.class,
  PS
)
}
\description{
\code{create.strat.flg.f.SKJ} yyy
}
