% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.strat.flg.f.YFT.R
\name{create.strat.flg.f.YFT}
\alias{create.strat.flg.f.YFT}
\title{xxx}
\usage{
create.strat.flg.f.YFT(
  lat.5deg,
  lon.5deg,
  is.lwrght,
  month,
  setype,
  vessel.class,
  PS
)
}
\description{
\code{create.strat.flg.f.YFT} yyy
}
