% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery.estimates.7599.f.R
\name{fishery.estimates.7599.f}
\alias{fishery.estimates.7599.f}
\title{xxx}
\usage{
fishery.estimates.7599.f(
  strat.ests.withsamps,
  spp.unlds.bystrat,
  iyear,
  PS,
  Species
)
}
\description{
\code{fishery.estimates.7599.f} yyy
}
