% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery.estimates.f.R
\name{fishery.estimates.f}
\alias{fishery.estimates.f}
\title{xxx}
\usage{
fishery.estimates.f(
  strat.ests.withsamps,
  strat.ests.nosamps,
  iyear,
  PS,
  Species
)
}
\description{
\code{fishery.estimates.f} yyy
}
