% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.catch.estimates.7599.f.R
\name{get.catch.estimates.7599.f}
\alias{get.catch.estimates.7599.f}
\title{xxx}
\usage{
get.catch.estimates.7599.f(
  cae.in,
  caestrtflg.in,
  totunlds.in,
  lfgrpd.in,
  lfmm.in,
  my.year,
  minsamps.in,
  well.estimates,
  PS,
  Species
)
}
\description{
\code{get.catch.estimates.7599.f} yyy
}
