% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.catch.estimates.f.R
\name{get.catch.estimates.f}
\alias{get.catch.estimates.f}
\title{xxx}
\usage{
get.catch.estimates.f(
  cae.in,
  caestrtflg.in,
  totunlds.in,
  lfgrpd.in,
  lfgrpd.stratflg.in,
  lfmm.in,
  my.year,
  minsamps.in,
  well.estimates,
  myarea.submat,
  growshrink.incrs.mat.touse,
  PS,
  Species,
  fishery.sub.mat
)
}
\description{
\code{get.catch.estimates.f} yyy
}
