% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.corrected.unloads.f.R
\name{get.corrected.unloads.f}
\alias{get.corrected.unloads.f}
\title{xxx}
\usage{
get.corrected.unloads.f(
  raw_data_dir,
  afile,
  start.year,
  end.year,
  corrections.frm
)
}
\description{
\code{get.corrected.unloads.f} yyy
}
