% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.strat.unloads.7599.f.R
\name{get.strat.unloads.7599.f}
\alias{get.strat.unloads.7599.f}
\title{xxx}
\usage{
get.strat.unloads.7599.f(cae.frm, cae.stratum, total.unloads)
}
\description{
\code{get.strat.unloads.7599.f} yyy
}
