% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sub.f.R
\name{get.sub.f}
\alias{get.sub.f}
\title{xxx}
\usage{
get.sub.f(
  area.str,
  mon.str,
  gear.str,
  fishery.areagear,
  lfgrp.stratflg,
  gear.submat,
  area.submat,
  month.submat,
  min.subsize,
  lfgrp.sampleid
)
}
\description{
\code{get.sub.f} yyy
}
