% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.summedLF.5deg.7599.YFT.f.R
\name{get.summedLF.5deg.7599.YFT.f}
\alias{get.summedLF.5deg.7599.YFT.f}
\title{xxx}
\usage{
get.summedLF.5deg.7599.YFT.f(well.samps, dph.vesno.flg, nyear)
}
\description{
\code{get.summedLF.5deg.7599.YFT.f} yyy
}
