% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.VB.output.new.V2.f.R
\name{get.VB.output.new.V2.f}
\alias{get.VB.output.new.V2.f}
\title{xxx}
\usage{
get.VB.output.new.V2.f(
  dir,
  name.unloads,
  name.cae,
  name.lfmm,
  name.lfgrpd,
  yr.start,
  yr.end
)
}
\description{
\code{get.VB.output.new.V2.f} yyy
}
