\name{gh.all}
\alias{gh.all}
\title{gh.all - Get likelihood for total number of observed patterns }
\description{
  Get likelihood for total number of observed patterns
}
\usage{
gh.all(x, amatrix, bmatrix, k, npoints)
}

\arguments{
  \item{x}{ ... }
  \item{amatrix}{ ... }
  \item{bmatrix}{ ... }
  \item{k}{ ... }
  \item{npoints}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
