\name{gh.each}
\alias{gh.each}
\title{gh.each - Approximate each integral using G-H }
\description{
  Approximate each integral using G-H
}
\usage{
  gh.each(x, a, b, k, npoints)
}
\arguments{
  \item{x}{ ... }
  \item{a}{ ... }
  \item{b}{ ... }
  \item{k}{ ... }
  \item{npoints}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
