\name{input66002}
\alias{input66002}
\docType{data}

\title{
  Example data.
}

\description{
  Example data.
}
\usage{
  data(input66002)
}
\format{
  TODO
}
\details{
  TODO
}
\source{
  TODO
}
\references{
  TODO
}
\examples{
  library(BSE)
  data(input66002)   # Load the dataset
}
\keyword{datasets}
