\name{margGH.bs}
\alias{margGH.bs}
\title{margGH.bs - Create the reference function g(theta) and computes the BS marginal likelihood estimators }
\description{
  Create the reference function g(theta) and computes the BS marginal likelihood estimators
}
\usage{
  margGH.bs(y, k, amatrix, bmatrix, npoints)
}
\arguments{
  \item{y}{ ... }
  \item{k}{ ... }
  \item{amatrix}{ ... }
  \item{bmatrix}{ ... }
  \item{npoints}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
