\name{meanlog}
\alias{meanlog}
\title{meanlog - De-log large numbers }
\description{
  De-log large numbers
}
\usage{
  meanlog(x)
}
\arguments{
  \item{x}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
