\name{metropolis}
\alias{metropolis}
\title{metropolis - Parameter estimation }
\description{
  Parameter estimation: Fit the model in winbugs
}
\usage{
metropolis(y, k, n.iter = 600, n.burnin = 100, n.thin = 1)
}

\arguments{
  \item{y}{ ... }
  \item{k}{ ... }
  \item{n.iter}{ ... }
  \item{n.burnin}{ ... }
  \item{n.thin}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
