\name{obsfreq.mp}
\alias{obsfreq.mp}
\title{obsfreq.mp - Get observed patterns }
\description{
  Get observed patterns
}
\usage{
  obsfreq.mp(x)
}
\arguments{
  \item{x}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
