\name{postmat}
\alias{postmat}
\title{postmat - Read Metropolis output }
\description{
  Read Metropolis output
}
\usage{
  postmat(eachtemp, s, k)
}
\arguments{
  \item{eachtemp}{ ... }
  \item{s}{ ... }
  \item{k}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
