% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.caedata.1deg.withvesscap.f.R
\name{read.caedata.1deg.withvesscap.f}
\alias{read.caedata.1deg.withvesscap.f}
\title{xxx}
\usage{
read.caedata.1deg.withvesscap.f(afile, yr.start, yr.end)
}
\description{
\code{read.caedata.1deg.withvesscap.f} yyy
}
