% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum.estimates.7599.f.R
\name{stratum.estimates.7599.f}
\alias{stratum.estimates.7599.f}
\title{xxx}
\usage{
stratum.estimates.7599.f(
  unloads.bystrat,
  well.estimates,
  well.strats,
  min.sampsize
)
}
\description{
\code{stratum.estimates.7599.f} yyy
}
