% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum.estimates.f.R
\name{stratum.estimates.f}
\alias{stratum.estimates.f}
\title{xxx}
\usage{
stratum.estimates.f(unloads.bystrat, well.estimates, well.strats, min.sampsize)
}
\description{
\code{stratum.estimates.f} yyy
}
