\name{BSE-package}
\alias{BSE}
\docType{package}
\title{
  BSE -  Bridge sampling estimators
}
\description{
  Description
}
\details{
  Details
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com}\cr
}
\references{
  \itemize{
    \item REF1
    \item REF2
  }
}
\seealso{
  \itemize{
    \item \code{\link{function1}} - Function 1
    \item \code{\link{function2}} - Function 2
  }
}
\keyword{package}
