\name{metropolis}
\alias{metropolis}
\title{metropolis - Get observed patterns }
\description{
  Parameter estimation: Fit the model in winbugs
}
\usage{
metropolis(y, k, R, Rburn, Rthin)
}

\arguments{
  \item{y}{ ... }
  \item{k}{ ... }
  \item{R}{ ... }
  \item{Rburn}{ ... }
  \item{Rthin}{ ... }
}
\details{
  TODO
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{ 
  Silia Vitoratou \email{vasiliki.vitoratou@kcl.ac.uk} and Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\seealso{
  TODO
}
\examples{
  #TODO
}
\keyword{methods}
