% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.f.R
\name{substitute.f}
\alias{substitute.f}
\title{xxx}
\usage{
substitute.f(
  strat.defns,
  unloads.bystrat.miss,
  lfgrpd.all.frm,
  lfgrpd.stratflg,
  lfmm.all.frm,
  gear.sub,
  area.sub,
  month.sub,
  min.subsize,
  growth.increments.array,
  fo.fishery.submat,
  un.fishery.submat,
  dp.fishery.submat
)
}
\description{
\code{substitute.f} yyy
}
